refactor: rename circuitBreaker package to circuitbreaker and move files
